cleaned comments

+++ project/src/program/mod.rs
@@ -1,7 +1,7 @@
 extern crate libc;
 use std;
 use std::ffi::{CString};
-use libc::c_int;
+use libc::{WEXITED, c_int};
 use super::ptrace;
 
 
@@ -26,8 +26,6 @@ impl Program {
     // Methods
     pub fn add_args(&mut self, program_args: Vec<*const i8>) {
         // Prepare the argv
-        //self.args.push(self.executable);
-        //self.args.push(std::ptr::null());
 
         self.args.clear();
         self.args = program_args;
@@ -41,25 +39,17 @@ impl Program {
             let cprogram_str = CString::new((self.executable).clone()).unwrap();
             let cprogram = cprogram_str.as_ptr();
 
-            // Prepare the environment
-            // let mut vec_environment: Vec<*const i8> = Vec::new();
-            // vec_environment.push(CString::new("HOME=~/Desktop").expect("vec_env_push").as_ptr());
-            // vec_environment.push(std::ptr::null());
-
-            //for x in &self.args {
-            //    println!("{:?}", **x);
-            //}
             // Start the executable
-            let ret = libc::execv(cprogram, self.args.as_ptr()/* , vec_environment.as_ptr() */);
+            let ret = libc::execv(cprogram, self.args.as_ptr());
             println!("Return was ............. {:?}", ret);
 
-            // Failed to run
+            // If failed to run
             println!("[ERROR] Failed to run, exited with err {:?} and errno {}", ret, *libc::__errno_location());
         }
     }
 
 
-    pub fn wait(&mut self) -> u32 {
+    pub fn wait(&mut self) -> i32 {
         let mut status: i32 = 0;
 
         unsafe {
@@ -67,6 +57,6 @@ impl Program {
             libc::waitpid(-1, &mut status, 0);
         }
 
-        return status as u32;
+        return status; 
     }
 }
